CDP #46 - Adding utils/search to move facet label logic to common location
