+++ package.json
@@ -27,9 +27,9 @@
     "@nanostores/react": "^0.7.1",
     "@netlify/functions": "^2.8.2",
     "@octokit/rest": "^21.0.2",
-    "@performant-software/core-data": "2.2.24-beta.30",
-    "@performant-software/geospatial": "2.2.24-beta.30",
-    "@performant-software/shared-components": "2.2.24-beta.30",
+    "@performant-software/core-data": "2.2.24-beta.32",
+    "@performant-software/geospatial": "2.2.24-beta.32",
+    "@performant-software/shared-components": "2.2.24-beta.32",
     "@peripleo/maplibre": "^0.5.2",
     "@peripleo/peripleo": "^0.5.2",
     "@radix-ui/react-dialog": "^1.0.5",
