+++ src/apps/search/Facet.tsx
@@ -1,5 +1,5 @@
 import TranslationContext from '@apps/search/TranslationContext';
-import { Typesense as TypesenseUtils } from '@performant-software/core-data';
+import { getFacetLabel } from '@utils/search';
 import clsx from 'clsx';
 import { useContext, useMemo, type ReactNode } from 'react';
 
@@ -15,20 +15,7 @@ const Facet = ({ attribute, children, className }: Props) => {
   /**
    * Memo-izes the label for the current facet.
    */
-  const label = useMemo(() => {
-    let value;
-
-    const relationshipId = TypesenseUtils.getRelationshipId(attribute);
-    const fieldId = TypesenseUtils.getFieldId(attribute);
-
-    if (relationshipId && fieldId) {
-      value = t('facetLabel', { relationship: t(relationshipId), field: t(fieldId) })
-    } else if (fieldId) {
-      value = t(fieldId);
-    }
-
-    return value;
-  }, [attribute, t]);
+  const label = useMemo(() => getFacetLabel(attribute, t), [attribute, t]);
 
   return (
     <div
